Feat/pytorch perf improvements (#545)
* set num_workers as a parameter

* write down a few performance tuning recommendations

* don't try to be smart about num_workers

* improved dtype support, fix a TFT issue

* datasets in float32 by default

* use self.__class__ where possible

* revert to default np.float64 for simplicity

* add simple benchmark performance

* added a small benchmarking example

* Update darts/models/forecasting/torch_forecasting_model.py


* Update darts/models/forecasting/torch_forecasting_model.py


* Update guides/performance.md

+++ darts/models/forecasting/tft_model.py
@@ -825,7 +825,11 @@ class TFTModel(TorchParametricProbabilisticForecastingModel, MixedCovariatesTorc
                                              freq=ts.freq) for ts in target]
 
         encoded_times = [
-            datetime_attribute_timeseries(ts, attribute=self.add_cyclic_encoder, cyclic=True) for ts in encode_ts
+            datetime_attribute_timeseries(ts, 
+                                          attribute=self.add_cyclic_encoder, 
+                                          cyclic=True, 
+                                          dtype=target[0].dtype) 
+            for ts in encode_ts
         ]
 
         if future_covariates is None:
